docs(tools): add commitzen and pre-commit
Commitzen is a helper cli which works like fzf helping you to choose the commit verb
and guides you in writing the commit message instead of using nano or any other editor
Pre-Commit is a Git Hook Manager which helps you to install pre-hooks in a programmatic way.
Currently the use case is to also have a final check for commit messages to confirm if they adhere
with conventional-commits format.