feat(helm): proper boolean check with false positive example
Using if conditional directly with value access is not guaranteed to work properly. It will usually throw a false positive such as in the case of using index function to get the value of a key which has a special character in it. In other cases it might work but the best way to do it is to use the dedicated function eq.